# api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Financial System API
  version: 1.0.0
  description: API REST para criacao de contas, processamento de operacoes financeiras e consulta de historico.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/operations/{type}:
    post:
      tags:
        - Operations
      summary: Processa uma operacao financeira
      description: Processa credit, debit, reserve, capture, transfer ou reversal. A API gera o reference_id automaticamente e o devolve na resposta.
      operationId: createOperation
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
            enum: [credit, debit, reserve, capture, transfer, reversal]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OperationRequest"
            examples:
              credit:
                value:
                  account_id: 1
                  value: 100000
                  currency: BRL
              transfer:
                value:
                  account_id: 1
                  account_dest_id: 2
                  value: 50000
                  currency: BRL
              reversal:
                value:
                  original_reference_id: 2d266f9f-01f8-44b0-a5ed-7f5da1a18d0c
      responses:
        "200":
          description: Operacao processada com sucesso.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/OperationSuccess"
                  - $ref: "#/components/schemas/TransferSuccess"
                  - $ref: "#/components/schemas/ReversalSuccess"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
...
  responses:
    BadRequest:
      description: Requisicao invalida.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflito de dados ou reference_id ja processado.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Recurso nao encontrado.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnprocessableEntity:
      description: Regra de negocio violada.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Erro interno.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [message, errorCode]
      properties:
        message:
          type: string
        errorCode:
          type: integer
...
    OperationRequest:
      type: object
      properties:
        account_id:
          type: integer
          format: uint64
          minimum: 1
        account_dest_id:
          type: integer
          format: uint64
          minimum: 1
        original_reference_id:
          type: string
          description: Obrigatorio para reversal.
        value:
          type: integer
          format: int64
          minimum: 1
          description: Valor em centavos.
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: BRL
    SuccessEnvelope:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    OperationData:
      type: object
      properties:
        transaction_id:
          type: string
        reference_id:
          type: string
          description: Identificador gerado pela API para idempotencia e estorno.
        status:
          type: string
          enum: [success, failed, pending]
        balance:
          type: integer
          format: int64
        reserved_balance:
          type: integer
          format: int64
        available_balance:
          type: integer
          format: int64
        credit_used:
          type: integer
          format: int64
        credit_available:
          type: integer
          format: int64
        timestamp:
          type: string
          format: date-time
        error_message:
          type: string
          nullable: true
        type:
          type: string
    OperationSuccess:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/OperationData"
    TransferData:
      type: object
      properties:
        transaction_id:
          type: string
        reference_id:
          type: string
          description: Identificador gerado pela API para idempotencia e estorno.
        status:
          type: string
        type:
          type: string
        origin_account_id:
          type: integer
          format: uint64
        dest_account_id:
          type: integer
          format: uint64
        origin_balance:
          type: integer
          format: int64
        dest_balance:
          type: integer
          format: int64
        credit_used:
          type: integer
          format: int64
        credit_available:
          type: integer
          format: int64
        timestamp:
          type: string
          format: date-time
        error_message:
          type: string
          nullable: true
        debit_tx_id:
          type: integer
          format: uint64
        credit_tx_id:
          type: integer
          format: uint64
        transfer_group_id:
          type: string
          nullable: true
    TransferSuccess:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/TransferData"
    ReversalAccountResult:
      type: object
      properties:
        account_id:
          type: integer
          format: uint64
        balance:
          type: integer
          format: int64
        available_balance:
          type: integer
          format: int64
        reserved_balance:
          type: integer
          format: int64
    ReversalData:
      type: object
      properties:
        transaction_id:
          type: string
        reference_id:
          type: string
          description: Identificador gerado pela API para idempotencia.
        status:
          type: string
        type:
          type: string
        timestamp:
          type: string
          format: date-time
        error_message:
          type: string
          nullable: true
        original_reference_id:
          type: string
        reversal_tx_ids:
          type: array
          items:
            type: integer
            format: uint64
        transfer_group_id:
          type: string
          nullable: true
        accounts:
          type: array
          items:
            $ref: "#/components/schemas/ReversalAccountResult"
    ReversalSuccess:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ReversalData"
...